JBTM-3430 Create a quarkus coordinator uber jar and remove JaxRsActivator (to enable services to be able to embed a coordinator)
